+++ src/plots/bullet/constant.ts
