+++ src/plots/histogram/constant.ts
