+++ src/plots/rose/constant.ts
